Add help text and email field
We build on the previous change adding the help text content expected to sit above the input field. We also add the label and input field, plus a matching attribute on the payment_summary_form. By setting the value of `:receipt_email` to `transient_registration.email_to_send_receipt` we get our pre-populated field which should handle the scenario of a user returning to this page having set a different email address.

Finally we have moved the section out of the 'multiple-choice' div. We lose the indentation. But the problem was our input was picking up the 'multiple-choice' class so was actually be set as invisible, as a small text box, and placed under the radio button!

The design for this was inspired by the 'Get a fishing licence service', specifically its 'How can we send you your licence details?'. If you select 'Email or text message' it displays a hidden section which contains an email field. Our section is displaying in the same way as it, with the lack of any indentation. So for now we're judging this as acceptable and moving on.