Refactor solution and add validation
Following the pattern that seems to be more recently used we delegate receipt_email to the transient_registration rather than creating our own attribute.

This means we can hook in the property getting saved and validated. We have also added `paying_by_card?` to handle the scenario of only validating the email if payment by card is selected.