Implement solution for prepopulating receipt_email
Using the new initialize callbacks built into `BaseForm` we call a method that will set the receipt email to whatever the transient_registration's `email_to_send_receipt()` logic tells us it should be. We expect 99.999% of the time it will be the current contact_email because receipt_email will not have been set. Though in 3 years time if the service is still running it will be when they come to renew 😁