Hide payment confirmation in back-office
https://eaflood.atlassian.net/browse/RUBY-1141

We have come across an issue where we need to change behaviour in the journey based on whether a registration or renewal is being carried out in either the front-office app or the back-office app.

We should have known, but have just realised that Worldpay does not send payment confirmation emails if the merchant code used is MOTO.

So there is no point in asking for and validating a payment confirmation email if the registration or renewal is created in the back-office.

So this change updates the engine to amend the behaviour of the payment summary screen, specifically whether to include the payment confirmation section, and expect a value in the POST based on whether the host app as told us it's the back-office.