Amend edit view to display payment confirmation
In the case of the edit view we'll always display the section even if its not populated. This mirrors how the account email section is handled.

Also added a spec for the edit_form_presenter, though just covering the new method I've added. We didn't seem to have one so a possible TODO to add coverage for the rest of the stuff in the future.