Remove dummy app .ruby-version file
Kept getting messages about using the wrong version of bundler when trying to run up the rails console in the dummy app.

Spotted that we had a .ruby-version file with the old version in it. I could have corrected the version but actually there is no need for it to be here when we have one in the root of the project.