Implement logic and tests
Documentation was scant but found there is a way to override the default getter Mongoid provides.

The key thing is to access the underlying data field you need to use `self[:field_name]`. So now we can just ask for `renew_token` where we need it, and one will be generated automatically if it doesn't exist.