Do not copy receipt_email when init. renewal
https://eaflood.atlassian.net/browse/RUBY-1172

This is the scenario

- User creates a new registration using the new journey in the FO
- User pays by card so a payment confirmation email is taken and :receipt_email is populated
- 3 years later user renews but as AD

Irrespective of how user pays, the renewed registration displays the payment confirmation section in Registration details

As we know that the payment confirmation email is only sent for digital users renewing or creating new registrations its a one-time thing against the registration.

So this means when a new renewal is initialised, we should not be copying the receipt email from the registration as it may not be applicable this time.