Hide contact name, phone and email in renewals
https://eaflood.atlassian.net/browse/RUBY-1171

NCCC want to allow as many users as possible to renew via magic link. 2 major blockers are

- 3rd party agents acting on behalf of users
- the 3 year gap between renewing where contacts at an organisation have changed

This means the automated renewal reminders containing magic links won't get to who they need to. When these people call NCCC to renew (because they didn't receive the reminder) NCCC would like to manually send them a registrations renewal link. We've already covered exposing the link in [PR #963](https://github.com/DEFRA/waste-carriers-back-office/pull/963). But to ensure we don't inadvertently expose any personal details NCCC would also like the contact details which are not currently public, to not pre-populate for renewals.

Ideally we would have 2 journeys; one for recognised users with as much information as possible pre-populated, and one for anonymous renewals. But with our current resources reducing the amount we pre-populate will help more people renew online more quickly in the short-term.

So this change stops the following being copied into the renewal when initialised from a registration

- contact name
- contact phone
- contact email