Allow for multiple grace window rules in engine
https://eaflood.atlassian.net/browse/RUBY-1210

Currently we have the COVID extension in place in the engine, with an additional ruleset overriding it in the back office if a feature flag is on.

We want to be able to continue to apply the COVID extension to registrations which previously had it, while applying the regular rules to any registrations expiring after the extension, and also still having the option for the feature-flagged ruleset to take effect as well.

This also includes substantial refactoring of the ExpiryCheckService as it was doing way too many things at once and splitting it up will make this much easier.