Use new grace window rules in ExpiryCheckService
This entire service is in serious need of breaking up into smaller chunks but this will have to be a start.

Had to make some changes to the specs as well, mainly to deal with there being two different grace windows depending on what date the tests are run. A new helper class should manage this for us.