Fix hardcoded links to old frontend
https://eaflood.atlassian.net/browse/RUBY-1223

We spotted that some pages (and an email) had hardcoded links to the old frontend. Since the frontend is being removed, we need to update these to go to the front office instead.

One of these instances was no longer being used in the code at all, so I have removed it.