Fix EPR query
This was including registrations where the registration's expires_on was less than or equal to the cutoff date. We actually want to exclude any regs from the cutoff date and onwards.

Unit tests didn't catch this because the expires_on was being set to later in the day on the cutoff date, rather than 00:00, so it was considered later than.