[RUBY-2845] Refactor certificate token generation and update services
This commit introduces a significant refactor to the way view certificate tokens are generated and handled within the application.
The primary goal of these changes is to improve the logic around the generation of new tokens and ensure that they are generated when necessary,
specifically so that emails will no longer be sent out without any token (and therefore a broken link).